fix(home): fix fragment context memory leak
Details: a memory leak was caused by having a combination of adapter click callbacks from FragmentWines & using a prepopulation technique for the recycled view pool.
This leak was causing our callback to be called in detached fragment, leading to crash on requireContext, since context is null and the fragemnt shoudl have been GCed but wasn't
