+++ app/src/main/java/com/louis/app/cavity/ui/home/FragmentWines.kt
@@ -7,17 +7,10 @@ import androidx.core.content.ContextCompat
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.navigation.fragment.FragmentNavigatorExtras
-import androidx.navigation.fragment.findNavController
-import androidx.transition.Transition
-import com.google.android.material.transition.MaterialSharedAxis
 import com.louis.app.cavity.R
 import com.louis.app.cavity.databinding.FragmentWinesBinding
-import com.louis.app.cavity.model.Wine
-import com.louis.app.cavity.util.L
 import com.louis.app.cavity.util.TransitionHelper
 import com.louis.app.cavity.util.setVisible
-import com.louis.app.cavity.util.toBoolean
 
 class FragmentWines : Fragment(R.layout.fragment_wines) {
     private var _binding: FragmentWinesBinding? = null
@@ -40,16 +33,7 @@ class FragmentWines : Fragment(R.layout.fragment_wines) {
 
         val wineAdapter = WineRecyclerAdapter(
             icons,
-            onItemClick = { wine, bottles, itemView ->
-                if (bottles.isNotEmpty()) {
-                    navigateToBottleDetails(wine, itemView)
-                } else {
-                    navigateToAddBottle(wine)
-                }
-            },
-            onItemLongClick = { wine, _ ->
-                navigateToWineOptionsBottomSheet(wine)
-            }
+            TransitionHelper(requireParentFragment())
         ).apply {
             setHasStableIds(true)
         }
@@ -122,46 +106,6 @@ class FragmentWines : Fragment(R.layout.fragment_wines) {
 //                override fun onTransitionPause(transition: Transition) = Unit
 //                override fun onTransitionResume(transition: Transition) = Unit
 //            })
-    }
-
-    private fun navigateToBottleDetails(wine: Wine, itemView: View) {
-        TransitionHelper(requireParentFragment()).setElevationScale()
-        recyleViewsOnExit()
-
-        val transition =
-            requireContext().getString(R.string.transition_bottle_details, wine.id)
-        val extra = FragmentNavigatorExtras(itemView to transition)
-        val action = FragmentHomeDirections.homeToBottleDetails(wine.id, -1)
-        findNavController().navigate(action, extra)
-    }
-
-    private fun navigateToAddBottle(wine: Wine) {
-        TransitionHelper(requireParentFragment()).setSharedAxisTransition(
-            MaterialSharedAxis.Z,
-            true
-        )
-        recyleViewsOnExit()
-
-        val action = FragmentHomeDirections.homeToAddBottle(wine.id, -1L)
-        findNavController().navigate(action)
-    }
-
-    private fun navigateToWineOptionsBottomSheet(wine: Wine) {
-        TransitionHelper(requireParentFragment()).setSharedAxisTransition(
-            MaterialSharedAxis.Z,
-            navigatingForward = true
-        )
-        recyleViewsOnExit()
-
-        val action = FragmentHomeDirections.homeToWineOptions(
-            wine.id,
-            wine.countyId,
-            wine.name,
-            wine.naming,
-            wine.isOrganic.toBoolean(),
-            wine.color
-        )
-        findNavController().navigate(action)
     }
 
     override fun onPause() {
