+++ app/src/main/java/com/louis/app/cavity/ui/home/WineRecyclerAdapter.kt
@@ -2,26 +2,23 @@ package com.louis.app.cavity.ui.home
 
 import android.graphics.drawable.Drawable
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.louis.app.cavity.R
 import com.louis.app.cavity.databinding.ItemWineBinding
 import com.louis.app.cavity.db.dao.WineWithBottles
-import com.louis.app.cavity.model.Bottle
-import com.louis.app.cavity.model.Wine
+import com.louis.app.cavity.util.TransitionHelper
 
 class WineRecyclerAdapter(
     private val drawables: Pair<Drawable, Drawable>,
-    private val onItemClick: (wine: Wine, bottles: List<Bottle>, itemView: View) -> Unit,
-    private val onItemLongClick: (wine: Wine, bottles: List<Bottle>) -> Unit
+    private val transitionHelper: TransitionHelper
 ) :
     ListAdapter<WineWithBottles, WineViewHolder>(WineItemDiffCallback()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): WineViewHolder {
         val binding = ItemWineBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return WineViewHolder(binding, drawables, onItemClick, onItemLongClick)
+        return WineViewHolder(binding, drawables, transitionHelper)
     }
 
     override fun onBindViewHolder(holder: WineViewHolder, position: Int) {
