+++ app/src/main/java/com/louis/app/cavity/ui/home/WineViewHolder.kt
@@ -4,24 +4,24 @@ import android.graphics.Color
 import android.graphics.drawable.Drawable
 import android.net.Uri
 import android.os.Build
-import android.view.View
 import androidx.core.content.ContextCompat
 import androidx.core.content.res.ResourcesCompat
 import androidx.core.view.ViewCompat
+import androidx.navigation.findNavController
+import androidx.navigation.fragment.FragmentNavigatorExtras
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.google.android.material.transition.MaterialSharedAxis
 import com.louis.app.cavity.R
 import com.louis.app.cavity.databinding.ItemWineBinding
 import com.louis.app.cavity.db.dao.WineWithBottles
-import com.louis.app.cavity.model.Bottle
-import com.louis.app.cavity.model.Wine
+import com.louis.app.cavity.util.TransitionHelper
 import com.louis.app.cavity.util.toBoolean
 
 class WineViewHolder(
     private val binding: ItemWineBinding,
     private val drawables: Pair<Drawable, Drawable>,
-    private val onItemClick: (wine: Wine, bottles: List<Bottle>, itemView: View) -> Unit,
-    private val onItemLongClick: (wine: Wine, bottles: List<Bottle>) -> Unit
+    private val transitionHelper: TransitionHelper
 ) :
     RecyclerView.ViewHolder(binding.root) {
 
@@ -65,11 +65,35 @@ class WineViewHolder(
         }
 
         itemView.setOnClickListener {
-            onItemClick(wine, bottles, binding.root)
+            if (bottles.isNotEmpty()) {
+                transitionHelper.setElevationScale()
+
+                val transition =
+                    itemView.context.getString(R.string.transition_bottle_details, wine.id)
+                val extra = FragmentNavigatorExtras(hexagone to transition)
+                val action = FragmentHomeDirections.homeToBottleDetails(wine.id, -1)
+                itemView.findNavController().navigate(action, extra)
+            } else {
+                transitionHelper.setSharedAxisTransition(MaterialSharedAxis.Z, true)
+
+                val action = FragmentHomeDirections.homeToAddBottle(wine.id, -1L)
+                itemView.findNavController().navigate(action)
+            }
         }
 
         itemView.setOnLongClickListener {
-            onItemLongClick(wine, bottles)
+            transitionHelper.setSharedAxisTransition(MaterialSharedAxis.Z, navigatingForward = true)
+
+            val action = FragmentHomeDirections.homeToWineOptions(
+                wine.id,
+                wine.countyId,
+                wine.name,
+                wine.naming,
+                wine.isOrganic.toBoolean(),
+                wine.color
+            )
+            itemView.findNavController().navigate(action)
+
             true
         }
     }
